Moving IoControlTransferType to Interop namespace, migrating usages
